backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6
supabase==2.0.2
python-dotenv>=1.0.0
httpx>=0.24.0,<0.25.0
google-generativeai>=0.3.0 

celery
redis 